+++ front/src/features/charts/utils/chart-data-transformer.ts
