--- front/src/features/search/constants/error-messages.ts
@@ -1,5 +1,0 @@
-export const ERROR_MESSAGES = {
-  API_ERROR: 'エラーが発生しました',
-  SEARCH_ERROR: '検索中にエラーが発生しました',
-  SEARCH_SUCCESS: '検索を実行しました',
-} as const;
