+++ front/src/features/search/utils/zod-rules.ts
