+++ front/src/features/universities/types/api-endpoints-paths.ts
