+++ front/src/features/universities/utils/university-finder.ts
